2020-11-22

+++ test/jdk/java/util/stream/boottest/java.base/java/util/stream/DoubleNodeTest.java
@@ -66,7 +66,7 @@ public class DoubleNodeTest extends OpTestCase {
     private static void assertEqualsListDoubleArray(List<Double> list, double[] array) {
         assertEquals(list.size(), array.length);
         for (int i = 0; i < array.length; i++)
-            assertEquals(array[i], list.get(i));
+            assertEquals(array[i], (double) list.get(i));
     }
 
     private List<Double> toList(double[] a) {
